Improve organization of fluent builder method renderers
